rename core headers to hpp, move to subfolder
