+++ src/Core/R4300/VCR.cpp
@@ -4,14 +4,14 @@
  * SPDX-License-Identifier: GPL-2.0-or-later
  */
 
-#include "core_types.h"
+#include "m64rr/core_types.hpp"
 #include <CommonPCH.hpp>
 #include <Core.hpp>
 #include <cassert>
 #include <Cheats.hpp>
 #include <filesystem>
 #include <format>
-#include <include/core_api.h>
+#include <m64rr/core_api.hpp>
 #include <iterator>
 #include <R4300/R4300.hpp>
 #include <R4300/Rom.hpp>
