+++ src/Core/include/m64rr/core_api.hpp
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "core_types.h"
+#include "m64rr/core_types.hpp"
 
 #ifdef __cplusplus
 extern "C"
