+++ src/Plugins.Win32.TASVideo/stdafx.h
@@ -2,7 +2,7 @@
 
 #include <CommonPCH.hpp>
 #include <VersionNameHelpers.hpp>
-#include <core_api.h>
+#include <m64rr/core_api.hpp>
 #include <Windows.h>
 #include <GL/glew.h>
 #include <commctrl.h>
